+++ src/engine/index.ts
@@ -5,3 +5,4 @@ export * from './Loop';
 export * from './DefaultAppDelegate';
 export * from './Model';
 export * from './ScriptHandler';
+export * from './DataHandler';