+++ src/interfaces/engine.ts
@@ -129,7 +129,7 @@ interface IExecution extends IServerComponent {
     error(msg: any): void;
     appendData(inputData: any,item:IItem, dataPath?: any,assignment?:any): void;
     getData(dataPath: any): any;
-    getAndCreateData(dataPath: any, asArray?: boolean): any;
+//    getAndCreateData(dataPath: any, asArray?: boolean): any;
 }
 
 interface IItem extends IItemData {